pyproject.toml
[tool.poetry]
name = "r2ditto"
version = "0.1.0"
description = "S3 Bucket based Image host"
authors = ["itssoap <admin@itssoap.ninja>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.29.6"
python-dotenv = "^1.0.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
fastapi = "^0.104.1"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
